ingestion/requirements.txt
# Ingestion Module Dependencies
# For standalone usage, install these packages
# For integrated usage, use root requirements.txt

# Core LLM and framework
langchain>=0.1.0
langchain-groq>=0.0.1
langchain-google-genai>=0.0.1
langgraph>=0.0.20

# Google Generative AI for image processing
google-generativeai>=0.3.0

# Document processing
PyPDF2>=3.0.0
pymupdf>=1.23.0  # fitz for advanced PDF
python-pptx>=0.6.21  # PowerPoint
python-docx>=0.8.11  # Word documents
docx2txt>=0.8  # Alternative Word processor

# Markdown and text processing
unstructured>=0.10.0
markdown>=3.4.0

# Image processing
Pillow>=10.0.0

# Data models
pydantic>=2.0.0

# Environment and utilities
python-dotenv>=1.0.0
typing-extensions>=4.5.0

# File type detection
filetype>=1.2.0

# LangSmith tracing
langsmith>=0.0.40

